Frisbee throw animation
Now character do a animation when throws the frisbee. It is applied with a animation mask, so it mixes with the other animations.
Also the character throws the frisbee with the animation.

Now climbing is faster, character positions corresponds with the animation and I added a little text to alert when can you climb.

Improved code.