Progress in tutorial and bugs
In the tutorial, after climbing a wall you must shoot a fire escape stairs to climb a building. Added the building too.
I created a script to play a animation when the collider of the object is hit by the frisbee.

Frisbee point system have been improved and its behaviour when it's more charged.
I have to reduce the speed of the frisbee in that situation...

Reduced bugs with the climbing system. The generated colliders are a little buggy. I added a new check that I think resolve a lot of this bugs.